chore: extract the arra-indexer CLI test fixture
- What: setup shared by all five describe blocks moves out of
  `arra-indexer.test.ts` (284 → 241) into `arra-indexer-harness.ts`. Allow-list
  entry removed, not lowered.

- Why: #2833, and CLAUDE.md's "nested, one behavior per file" — a 284-line file
  with five describe blocks and 45 lines of preamble is not that.

- Not named `*.test.ts` on purpose: bun collects by filename, so a fixture with
  that suffix runs as a suite containing no assertions — which passes, and reads
  as coverage.

The import is placed after the last line that CLOSES an import statement, not
after the last line that starts with `import`. That heuristic broke two files
earlier tonight by inserting into the middle of a multi-line import.

Gate: `bunx tsc --noEmit` exit 0 · `bun test --isolate src/indexer/__tests__/
tests/build/` 153 pass.

Allow-list: 22 at #2833's filing → **16**.

Refs #2833
